refactor(schema-config): extract store types and save helpers
