+++ specifyweb/frontend/js_src/lib/components/SchemaConfig/helpers.ts
@@ -1,8 +1,10 @@
 import { schemaText } from '../../localization/schema';
 import type { IR, RA } from '../../utils/types';
 import { localized } from '../../utils/types';
+import { replaceItem } from '../../utils/utils';
 import { addMissingFields } from '../DataModel/addMissingFields';
 import type { SerializedResource } from '../DataModel/helperTypes';
+import { createResource, saveResource } from '../DataModel/resource';
 import type { SpLocaleContainerItem } from '../DataModel/types';
 import type { Aggregator, Formatter } from '../Formatters/spec';
 import type {
@@ -11,6 +13,7 @@ import type {
   SpLocaleItemString,
 } from './index';
 import type { SchemaFormatter } from './schemaData';
+import type { SaveRequest, SchemaConfigEditorState } from './types';
 
 let newStringId = 1;
 const defaultLanguage = 'en';
@@ -101,3 +104,176 @@ export function javaTypeToHuman(
   else if (type.startsWith('java')) return type.split('.').at(-1)!;
   else return type;
 }
+
+const getEditorChanges = (
+  editor: SchemaConfigEditorState
+): {
+  readonly nameChanged: boolean;
+  readonly descChanged: boolean;
+  readonly containerChanged: boolean;
+} => ({
+  nameChanged:
+    JSON.stringify(editor.initialName) !== JSON.stringify(editor.name),
+  descChanged:
+    JSON.stringify(editor.initialDesc) !== JSON.stringify(editor.desc),
+  containerChanged:
+    JSON.stringify(editor.initialContainer) !==
+    JSON.stringify(editor.container),
+});
+
+export const isEditorModified = (editor: SchemaConfigEditorState): boolean => {
+  const changes = getEditorChanges(editor);
+  return (
+    changes.nameChanged ||
+    changes.descChanged ||
+    changes.containerChanged ||
+    editor.changedItems.length > 0
+  );
+};
+
+export const updateEditor = (
+  editors: IR<SchemaConfigEditorState>,
+  tableName: string,
+  update: (editor: SchemaConfigEditorState) => SchemaConfigEditorState
+): IR<SchemaConfigEditorState> => {
+  const editor = editors[tableName];
+  return editor === undefined
+    ? editors
+    : { ...editors, [tableName]: update(editor) };
+};
+
+const applySavedId = <T extends { readonly id?: number }>(
+  resource: T,
+  result: unknown
+): T =>
+  typeof (result as { readonly id?: number } | undefined)?.id === 'number'
+    ? { ...resource, ...(result as Partial<T>) }
+    : resource;
+
+const applyItemStringId = (
+  editor: SchemaConfigEditorState,
+  index: number,
+  key: 'name' | 'desc',
+  sent: NewSpLocaleItemString | SpLocaleItemString,
+  result: unknown
+): SchemaConfigEditorState => {
+  const item = editor.items[index];
+  const current = key === 'name' ? item.strings.name : item.strings.desc;
+  const saved = applySavedId(sent, result);
+  const string = current === sent ? saved : current;
+  const strings =
+    key === 'name'
+      ? { ...item.strings, name: string }
+      : { ...item.strings, desc: string };
+  return {
+    ...editor,
+    items: replaceItem(editor.items, index, { ...item, strings }),
+  };
+};
+
+export const buildSaveRequests = (
+  editor: SchemaConfigEditorState
+): RA<SaveRequest> => {
+  const { nameChanged, descChanged, containerChanged } =
+    getEditorChanges(editor);
+  // Capture the values sent with each request so reconciliation can compare
+  // the current state against what was actually saved
+  const sentName = editor.name;
+  const sentDesc = editor.desc;
+  const sentContainer = editor.container;
+  return [
+    ...(nameChanged
+      ? [
+          {
+            promise: saveString(sentName),
+            reconcile: (
+              current: SchemaConfigEditorState,
+              result: unknown
+            ): SchemaConfigEditorState => {
+              const saved = applySavedId(sentName, result);
+              return {
+                ...current,
+                name: current.name === sentName ? saved : current.name,
+                initialName: saved,
+              };
+            },
+          },
+        ]
+      : []),
+    ...(descChanged
+      ? [
+          {
+            promise: saveString(sentDesc),
+            reconcile: (
+              current: SchemaConfigEditorState,
+              result: unknown
+            ): SchemaConfigEditorState => {
+              const saved = applySavedId(sentDesc, result);
+              return {
+                ...current,
+                desc: current.desc === sentDesc ? saved : current.desc,
+                initialDesc: saved,
+              };
+            },
+          },
+        ]
+      : []),
+    ...(containerChanged
+      ? [
+          {
+            promise: saveResource(
+              'SpLocaleContainer',
+              sentContainer.id,
+              sentContainer
+            ),
+            reconcile: (
+              current: SchemaConfigEditorState
+            ): SchemaConfigEditorState => ({
+              ...current,
+              initialContainer: sentContainer,
+            }),
+          },
+        ]
+      : []),
+    ...editor.items.flatMap(({ strings, ...item }, index) =>
+      editor.changedItems.includes(index)
+        ? [
+            {
+              itemIndex: index,
+              promise: saveResource('SpLocaleContainerItem', item.id, item),
+              reconcile: (
+                current: SchemaConfigEditorState
+              ): SchemaConfigEditorState => current,
+            },
+            {
+              itemIndex: index,
+              promise: saveString(strings.name),
+              reconcile: (
+                current: SchemaConfigEditorState,
+                result: unknown
+              ): SchemaConfigEditorState =>
+                applyItemStringId(current, index, 'name', strings.name, result),
+            },
+            {
+              itemIndex: index,
+              promise: saveString(strings.desc),
+              reconcile: (
+                current: SchemaConfigEditorState,
+                result: unknown
+              ): SchemaConfigEditorState =>
+                applyItemStringId(current, index, 'desc', strings.desc, result),
+            },
+          ]
+        : []
+    ),
+  ];
+};
+
+const saveString = async (
+  resource: NewSpLocaleItemString | SpLocaleItemString
+): Promise<unknown> =>
+  'resource_uri' in resource &&
+  typeof resource.id === 'number' &&
+  resource.id >= 0
+    ? saveResource('SpLocaleItemStr', resource.id, resource)
+    : createResource('SpLocaleItemStr', resource);
